move `Snapshot` to submodule
